requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
cached-property==1.5.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
commonmark==0.9.1
dictdiffer==0.8.1
dominate==2.5.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mistletoe==0.7.2
notion-py==0.0.5
python-dotenv==0.14.0
python-slugify==4.0.1
pytz==2020.1
requests==2.24.0
soupsieve==2.0.1
text-unidecode==1.3
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
